requirements.txt
beautifulsoup4==4.14.2
Requests==2.32.5
discord.py